create ExtrinsicCard component
